+++ RELEASE_NOTES.md
@@ -1,8 +1,20 @@
-# Next Release
+# Next release
 
-- [#98](https://github.com/IAMconsortium/pyam/pull/986) Extend `set_meta_from_data()` to apply on different column
+## API changes
 
-# Release v3.3.3
+The method `Statistics.summarize()` includes a new argument `limits` that can take the
+values 'full range', 'interquartile' or 'central 90%'. The previous boolean arguments
+`fullrange` and `interquartile` are marked as deprecated.
+
+## Individual updates
+
+- [#993](https://github.com/IAMconsortium/pyam/pull/993) Mark `len(df)` as deprecated
+- [#991](https://github.com/IAMconsortium/pyam/pull/991) Support the central 90% as limits in the statistics summary
+- [#989](https://github.com/IAMconsortium/pyam/pull/989) Add an `IamDataFrame.series` attribute to get timeseries data
+  as **pd.Series**
+- [#986](https://github.com/IAMconsortium/pyam/pull/986) Extend `set_meta_from_data()` to apply on different column
+
+# Release v3.4.0
 
 ## Highlights